refractor: split interface and DAO
